# swagger_server/controllers/default_controller.py
import connexion
import six

from swagger_server.models.book import Book  # noqa: E501
from swagger_server.models.book_input import BookInput  # noqa: E501
from swagger_server.models.loan import Loan  # noqa: E501
from swagger_server.models.loan_input import LoanInput  # noqa: E501
from swagger_server.models.user import User  # noqa: E501
from swagger_server.models.user_input import UserInput  # noqa: E501
from swagger_server import util


def books_get():  # noqa: E501
    """Get a list of books

    Returns a list of all available books in the library. # noqa: E501


    :rtype: List[Book]
    """
    return 'do some magic!'


def books_id_delete(id):  # noqa: E501
    """Delete a book

    Deletes a book from the library by its ID. # noqa: E501

    :param id: 
    :type id: str

    :rtype: None
    """
    return 'do some magic!'


def books_id_get(id):  # noqa: E501
    """Get a book by ID

    Returns details of a specific book by its ID. # noqa: E501

    :param id: 
    :type id: str

    :rtype: Book
    """
    return 'do some magic!'


def books_id_put(body, id):  # noqa: E501
    """Update a book

    Updates the details of an existing book. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param id: 
    :type id: str

    :rtype: Book
    """
    if connexion.request.is_json:
        body = BookInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def books_post(body):  # noqa: E501
    """Add a new book

    Creates a new book record in the library. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Book
    """
    if connexion.request.is_json:
        body = BookInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def rental_get():  # noqa: E501
    """Get a list of rental

    Returns a list of all current book rental. # noqa: E501


    :rtype: List[Loan]
    """
    return 'do some magic!'


def rental_id_delete(id):  # noqa: E501
    """Delete a loan

    Deletes a loan from the system. # noqa: E501

    :param id: 
    :type id: str

    :rtype: None
    """
    return 'do some magic!'


def rental_id_get(id):  # noqa: E501
    """Get a loan by ID

    Returns details of a specific loan by its ID. # noqa: E501

    :param id: 
    :type id: str

    :rtype: Loan
    """
    return 'do some magic!'


def rental_id_put(body, id):  # noqa: E501
    """Update a loan

    Updates the details of an existing loan (e.g., return date, status). # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param id: 
    :type id: str

    :rtype: Loan
    """
    if connexion.request.is_json:
        body = LoanInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def rental_post(body):  # noqa: E501
    """Create a new loan

    Issues a new book loan to a user. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Loan
    """
    if connexion.request.is_json:
        body = LoanInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def users_get():  # noqa: E501
    """Get a list of users

    Returns a list of all registered users in the library system. # noqa: E501


    :rtype: List[User]
    """
    return 'do some magic!'


def users_id_delete(id):  # noqa: E501
    """Delete a user

    Deletes a user from the library system by their ID. # noqa: E501

    :param id: 
    :type id: str

    :rtype: None
    """
    return 'do some magic!'


def users_id_get(id):  # noqa: E501
    """Get a user by ID

    Returns details of a specific user by their ID. # noqa: E501

    :param id: 
    :type id: str

    :rtype: User
    """
    return 'do some magic!'


def users_id_put(body, id):  # noqa: E501
    """Update a user

    Updates the details of an existing user. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param id: 
    :type id: str

    :rtype: User
    """
    if connexion.request.is_json:
        body = UserInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def users_post(body):  # noqa: E501
    """Register a new user

    Creates a new user in the library system. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: User
    """
    if connexion.request.is_json:
        body = UserInput.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'